Changed screen size to 16x16 Pixels
Changing the screen size to 16x16 pixels to make sure
the SNAKE-buffers cannot overflow. The SNAKE-buffers are two 256 byte
buffers for x- and y-coords of the different snake-elements. They are
by x-indexed accessing. By reducing the screen size, buffer overflows
are prevented.

The screen size is reduced by drawing a border around the snake.
The border is made of snake characters, to reuse the collision detection.
That way, old border collision detection can be removed and move_head
is turned back into a proper subroutine.
Fruit spawning logic is also modified to force fruits to spawn inside
the border.

Additionally, cosmetic changes include the addition of the title
"SNAKE GAME!"